+++ mycodo/scripts/pandoc_generate.sh
@@ -2,19 +2,21 @@
 # Generates the Mycodo Manual in different formats with pandoc
 #
 # To install dependencies on Debian/Ubuntu:
-# sudo apt-get install pandoc texlive texlive-xetex
+# sudo apt-get install pandoc
 #
 
 INSTALL_DIRECTORY=$( cd "$( dirname "${BASH_SOURCE[0]}" )/../../" && pwd -P )
+cd ${INSTALL_DIRECTORY}
 
 # check if pandoc is installed
 if [[ `command -v pandoc` ]]
 then
-  # output PDF file
-  pandoc ${INSTALL_DIRECTORY}/mycodo/scripts/pandoc_metadata.yaml ${INSTALL_DIRECTORY}/mycodo-manual.md --latex-engine=xelatex -s -o ${INSTALL_DIRECTORY}/mycodo-manual.pdf
+  # output PDF file (dep: sudo apt-get install texlive texlive-xetex)
+  # disabled because it doesn't render images correctly
+  # pandoc ${INSTALL_DIRECTORY}/mycodo/scripts/pandoc_metadata.yaml ${INSTALL_DIRECTORY}/mycodo-manual.md --latex-engine=xelatex -s -o ${INSTALL_DIRECTORY}/mycodo-manual.pdf
 
   # output HTML (HTML5) file
-  pandoc ${INSTALL_DIRECTORY}/mycodo-manual.md -H ${INSTALL_DIRECTORY}/mycodo/scripts/pandoc.css_style -s -S -t html5 -o ${INSTALL_DIRECTORY}/mycodo-manual.html
+  pandoc ${INSTALL_DIRECTORY}/mycodo-manual.md -H ${INSTALL_DIRECTORY}/mycodo/scripts/pandoc.css_style --self-contained -s -S -t html5 -o ${INSTALL_DIRECTORY}/mycodo-manual.html
   cp ${INSTALL_DIRECTORY}/mycodo-manual.html ${INSTALL_DIRECTORY}/mycodo/mycodo_flask/templates/manual.html
 
   # output plain text